Update README.md
Are you looking for a trusted crypto exchange that offers generous rewards, low trading fees, and a seamless user experience? Bitget is the platform to watch in 2025! And with the exclusive Bitget referral code: **BEST99**, new users can earn up to **15,000 USDT** in welcome bonuses and enjoy a **50% discount** on trading fees.